feat : collect dianostics from simulation
Once the phase-space is run, this change introduces a way to
immediately post-process the simulation to write dianostic data
out to file. For now a function that calulcates the steer and pose
rates and angles is provided. Future commits will aim at making the
statistics more general (don't harcode anything)