:bug:fix : calculation of circle fit
With straight-line or near straight-line data, the fitting is not
robust (it tries to only minimize the least squared error). So before
fitting, we first run a linear-least squares to fit to a line (the
edge case of a circle with infinite radius), see if that fails to fit
and only then do the circle.