Added computation of average trajectory over multiple periods.
Three consequetive periods are used to compute the average COM over each period. These points are then uesd to define a circle with radius R. From here the steering rate along this circle trajectory as well as an effective speed along the trajectory are computed.

THere are also a few small changes to run_and_visualize to plot the COM and avg COM.